refac: move file path to specific router separated by method
