+++ backend/main.py
@@ -1,13 +1,7 @@
-from fastapi import FastAPI, File, UploadFile, status, HTTPException
-import duckdb
-import io
+from fastapi import FastAPI
+from routes.file import post, get
 import uvicorn
 from fastapi.middleware.cors import CORSMiddleware
-import os
-from database.models.file.file import File as FileSchema
-from database.connection import DatabaseConnection
-from database.query.file.file import FileQuery
-from utils.local_exception import LocalException
 
 app = FastAPI()
 
@@ -21,86 +15,8 @@ app.add_middleware(
     allow_headers=["*"],
 )
 
-
-@app.post("/file/upload")
-async def post_file(file: UploadFile = File(...), separator: str = ',', header: bool = True):
-    fileContent = file.file.read()
-    try:
-        fileContentStr = fileContent.decode('utf-8')
-    except UnicodeDecodeError:
-        fileContentStr = fileContent.decode('ISO-8859-1')
-    fileName = file.filename.split(".")[0].lower().replace(" ", "_")
-    fileExtension = file.filename.split(".")[1].lower()
-    fileContentIo = io.StringIO(fileContentStr)
-
-    main = duckdb.read_csv(fileContentIo, sep=separator, header=header)
-    duckdb.sql(f"copy main to 's3/{file.filename}'")
-
-    path = "s3/"
-    files = os.listdir(path)
-    filesSizes = {file: f"{os.path.getsize(os.path.join(path, file)) / (1024 * 1024):.2f}" for file in files}
-
-    rowsNumber = duckdb.sql("SELECT COUNT(*) FROM main").fetchdf().iloc[0, 0]
-    dfColumns = list(duckdb.sql("SELECT * FROM main limit 1").to_df().columns)
-    fileSize = filesSizes.get(file.filename)
-
-    dbConnection = DatabaseConnection()
-    newFile = FileSchema(name=fileName,
-                         type=fileExtension,
-                         size=float(fileSize),
-                         nb_columns=len(dfColumns),
-                         nb_rows=int(rowsNumber)
-                         )
-
-    schemaInfo = duckdb.sql(f"DESCRIBE TABLE main").to_df()
-
-    insertFileResult = dbConnection.addNewObject(newFile)
-    if isinstance(insertFileResult, LocalException):
-        raise HTTPException(insertFileResult.status_code, insertFileResult.error_message)
-
-    fileQuery = FileQuery()
-    fileId = fileQuery.getFileByName(fileName).id
-
-
-    schemaInfoList = []
-    for column in dfColumns:
-        rowOfDfSchema = schemaInfo[schemaInfo['column_name'] == column]
-        columnType = rowOfDfSchema['column_type'].iloc[0]
-        hvNull = True if rowOfDfSchema['null'].iloc[0] == 'YES' else False
-        hvDefault = rowOfDfSchema['default'].iloc[0]
-        schemaInfoList.append({"id_table": fileId, "column_name": column, "column_type": columnType, "null": hvNull,
-                               "default": hvDefault})
-    schemaInsert = fileQuery.bulkInsertSchema(schemaInfoList)
-    if isinstance(schemaInsert, LocalException):
-        fileQuery.deleteFileByName(fileId)
-        raise HTTPException(schemaInsert.status_code, schemaInsert.error_message)
-
-    return status.HTTP_201_CREATED
-
-
-@app.get("/file/schema")
-async def get_schema(id: int):
-    fileQuery = FileQuery()
-    return {"schema": fileQuery.getSchemaByIdFile(id)}
-
-
-@app.post("/file/preview")
-async def file_preview(file: UploadFile = File(...), separator: str = ',', header: bool = True):
-    fileContent = await file.read()
-    fileContentStr = fileContent.decode('utf-8')
-    fileContentIo = io.StringIO(fileContentStr)
-
-    main = duckdb.read_csv(fileContentIo, sep=separator, header=header)
-    view = duckdb.sql("select * from main limit 5").to_df().to_dict(orient='records')
-
-    return view
-
-@app.get("/file/list")
-async def get_file_list():
-    fileQuerys = FileQuery()
-
-    return {"files": fileQuerys.getAllFiles()}
-
+app.include_router(post.router)
+app.include_router(get.router)
 
 if __name__ == "__main__":
     uvicorn.run("main:app", host="0.0.0.0", port=8000, log_level="info", reload=True)
